requirements.txt
requests
python-dotenv
cryptography
pyotp
qrcode[pil]
pandas